requirements.txt
viam-sdk~=0.21.0rc1
roboflow
docker
numpy